Added CellIterator Class as base for RadialCellIterator & SquareCellIterator this reduces duplicate code and should make the code more maintainable
